+++ src/Events.js
@@ -1,0 +1,24 @@
+import React from "react";
+import Event from "./Event";
+
+const Events = ({ state, dispatch }) => {
+  return (
+    <>
+      <h4>イベント一覧</h4>
+      <table className="table table-hover">
+        <thead>
+          <th>ID</th>
+          <th>タイトル</th>
+          <th>ボディー</th>
+        </thead>
+        <tbody>
+          {state.map((event, index) => (
+            <Event key={index} event={event} dispatch={dispatch} />
+          ))}
+        </tbody>
+      </table>
+    </>
+  );
+};
+
+export default Events;
